+++ src/Xamarin.Forms.Labs/Plugins/Charting/Xamarin.Forms.Labs.Charting/Events/SingleDrawingData.cs
@@ -9,11 +9,11 @@ namespace Xamarin.Forms.Labs.Charting.Events
     public sealed class SingleDrawingData
     {
         public int SeriesNo { get; set; }
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float Size { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Size { get; set; }
 
-        public SingleDrawingData(float x, float y, int seriesNo)
+        public SingleDrawingData(double x, double y, int seriesNo)
         {
             X = x;
             Y = y;
